requirements.txt
flask
requests
markdown
torch
transformers
python-dotenv
torch
groq
redis
pillow
google-search-results
flask-session